todo_list/requirements.txt
asgiref==3.3.4
cmsplugin-markdownx==0.1.3
Django==3.2.3
django-classy-tags==0.8.0
django-cms==3.4.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-formtools==2.1
django-markdownx==1.7.1
django-sekizai==0.10.0
django-treebeard==4.3.1
djangocms-admin-style==2.0.2
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
Faker==8.6.0
importlib-metadata==4.0.1
Markdown==3.3.4
mixer==7.1.2
Pillow==8.2.0
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
zipp==3.4.1
